pyproject.toml
[project]
name = "mbgv2-dissertation"
version = "0.1.0"
description = "MBGv2 dataset pre-processing and finetuning scripts for YOLO models."
authors = [{ name = "milasd", email = "mila.oliveira@coppe.ufrj.br" }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "ultralytics (==8.2.91)",
    "ruff (>=0.11.10,<0.12.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "torch (==2.3.1)",
    "torchvision (==0.18.1)",
    "sahi==0.11.19",
    "types-tqdm>=4.67.0.20250516",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-cov>=6.1.1,<7",
    "pandas-stubs",
]

[tool.uv]

[tool.hatch.build.targets.sdist]
include = ["MBGv2_dissertation"]

[tool.hatch.build.targets.wheel]
include = ["MBGv2_dissertation"]
